+++ services/app-api/src/testHelpers/rateDataMocks.ts
@@ -1,12 +1,12 @@
 import { v4 as uuidv4 } from 'uuid'
-import type { InsertRateArgsType } from '../../postgres/contractAndRates/insertRate'
+import type { InsertRateArgsType } from '../postgres/contractAndRates/insertRate'
 import type {
     RateTableFullPayload,
     RateRevisionTableWithContracts,
-} from '../../postgres/contractAndRates/prismaSubmittedRateHelpers'
-import type { StateCodeType } from '../../../../app-web/src/common-code/healthPlanFormDataType'
-import { findStatePrograms } from '../../postgres'
-import { must } from '../assertionHelpers'
+} from '../postgres/contractAndRates/prismaSubmittedRateHelpers'
+import type { StateCodeType } from 'app-web/src/common-code/healthPlanFormDataType'
+import { findStatePrograms } from '../postgres'
+import { must } from './assertionHelpers'
 
 const defaultRateData = () => ({
     id: '24fb2a5f-6d0d-4e26-9906-4de28927c882',
@@ -134,8 +134,4 @@ const mockRateRevision = (
     }
 }
 
-export {
-    mockInsertRateArgs,
-    mockRateRevision,
-    mockDraftRate,
-}
+export { mockInsertRateArgs, mockRateRevision, mockDraftRate }
